pyproject.toml
[project]
name = "crypto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.5",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
